+++ kernel/src/lib.rs
@@ -12,6 +12,7 @@ pub mod console;
 pub mod dtb;
 pub mod elf;
 pub mod memory;
+pub mod process_image;
 pub mod user;
 
 pub mod arch;